refactor: replace old rollup by new one
