+++ packages/rooks/rollup.config.ts
@@ -1,16 +1,41 @@
-import nodeResolve from "@rollup/plugin-node-resolve";
-import replace from "@rollup/plugin-replace";
-import commonjs from "@rollup/plugin-commonjs";
-import babel from "@rollup/plugin-babel";
-import json from "@rollup/plugin-json";
-import typescript from "rollup-plugin-typescript";
-import { terser } from "rollup-plugin-terser";
-import sourceMaps from "rollup-plugin-sourcemaps";
+import dts from "rollup-plugin-dts";
+import typescript2 from "rollup-plugin-typescript2";
+import { typescriptPaths } from "rollup-plugin-typescript-paths";
 import pkg from "./package.json";
+import nodeResolve from "@rollup/plugin-node-resolve";
+import sourceMaps from "rollup-plugin-sourcemaps";
+import json from "@rollup/plugin-json";
+import babel from "@rollup/plugin-babel";
+import commonjs from "@rollup/plugin-commonjs";
+import replace from "@rollup/plugin-replace";
+import { terser } from "rollup-plugin-terser";
 
-const commonPlugins = [
+const external = [
+  ...Object.keys(pkg.dependencies),
+  ...Object.keys(pkg.peerDependencies || {}),
+];
+
+const umdGlobals = { react: "React", "react-dom": "ReactDOM" };
+
+const distDir = "./dist/future";
+
+const tsPathPlugin = typescriptPaths({
+  preserveExtensions: true,
+  tsConfigPath: "./tsconfig.json",
+});
+
+const plugins = [
+  tsPathPlugin,
+  typescript2({
+    tsconfig: "./tsconfig.build.json",
+  }),
+];
+
+const umdPlugins = [
   nodeResolve(),
-  typescript(),
+  typescript2({
+    tsconfig: "./tsconfig.build.json",
+  }),
   sourceMaps(),
   json(),
   babel({
@@ -27,25 +52,62 @@ const commonPlugins = [
   }),
 ];
 
-const prodPlugins = [terser()];
+const prodUmdPlugins = [terser()];
 
-const globals = { react: "React", "react-dom": "ReactDOM" };
-
-const config = {
-  input: "./src/index.ts",
-
-  // \0 is rollup convention for generated in memory modules
-  external: Object.keys(globals),
-  plugins: commonPlugins,
-  output: [
-    {
-      file: "dist/umd/rooks.umd.js",
-      format: "umd",
-      globals,
-      name: "rooks",
-      plugins: prodPlugins,
+export default () => [
+  // ESM
+  {
+    input: ["./src/index.ts"],
+    output: {
+      dir: `${distDir}/esm`,
+      format: "esm",
+      sourcemap: false,
     },
-  ],
-};
-
-export default config;
+    preserveModules: true,
+    external,
+    plugins,
+  },
+  // CJS
+  {
+    input: ["./src/index.ts"],
+    output: {
+      dir: `${distDir}/cjs`,
+      format: "cjs",
+      sourcemap: false,
+    },
+    preserveModules: true,
+    external,
+    plugins: [...plugins],
+  },
+  {
+    input: "./src/index.ts",
+    output: {
+      file: `${distDir}/types/index.d.ts`,
+    },
+    external,
+    plugins: [
+      tsPathPlugin,
+      dts({
+        compilerOptions: {
+          sourceMap: false,
+        },
+      }),
+    ],
+  },
+  // @todo UMD from rollup.config.ts
+  {
+    input: "./src/index.ts",
+    // \0 is rollup convention for generated in memory modules
+    external: Object.keys(umdGlobals),
+    plugins: umdPlugins,
+    output: [
+      {
+        file: `${distDir}/umd/rooks.umd.js`,
+        format: "umd",
+        globals: umdGlobals,
+        name: "rooks",
+        plugins: prodUmdPlugins,
+      },
+    ],
+  },
+];
